+++ src/compiler/compiler-test/src/src_str.rs
@@ -7,37 +7,29 @@ d1 {
 
 kv MyKv {
     someValue(id1: string, id2: int) -> json {
-        "{id1}/{id2}"
-    }
-
-    manyValues() -> paginated<json> {
-        ""
+        "value/{id1}/{id2}"
     }
 
     streamValue(id1: string, id2: int) -> stream {
-        "{id1}/{id2}"
+        "stream/{id1}/{id2}"
     }
 }
 
 r2 MyR2 {
     fileData(id: string) {
-        "{id}"
-    }
-
-    manyFileDatas(id: string) paginated {
-        "{id}/files"
+        "files/{id}"
     }
 
     metadata(ownerId: string, modelYear: int) {
-        "{ownerId}/{modelYear}"
+        "meta/{ownerId}/{modelYear}"
     }
 
     photoData(modelYear: int) {
-        "{modelYear}/photos"
+        "photos/{modelYear}"
     }
 
     customDsData(id: int) {
-        "{id}/data"
+        "custom/{id}/data"
     }
 }
 
@@ -146,10 +138,6 @@ model ModelWithKv for db {
         someValue
     }
 
-    kv MyKv::manyValues() {
-        manyValues
-    }
-
     kv MyKv::streamValue(id1, id2) {
         streamValue
     }
@@ -169,10 +157,6 @@ model ModelWithR2 for db {
 
     r2 MyR2::fileData(id) {
         fileData
-    }
-
-    r2 MyR2::manyFileDatas(id) {
-        manyFileDatas
     }
 }